requirements.txt
requests
pyyaml
pefile
beautifulsoup4
